# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: JEE Test Mock Free API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/send-otp:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send OTP to phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpInput"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP and login or register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpInput"
      responses:
        "200":
          description: Auth successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /tests:
    get:
      operationId: listTests
      tags: [tests]
      summary: List all available tests
      parameters:
        - in: query
          name: subject
          schema:
            type: string
          description: Filter by subject
        - in: query
          name: type
          schema:
            type: string
          description: Filter by type (full/subject)
      responses:
        "200":
          description: List of tests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Test"
    post:
      operationId: createTest
      tags: [tests]
      summary: Create a new test (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Test"
  /tests/{testId}:
    get:
      operationId: getTest
      tags: [tests]
      summary: Get test details
      parameters:
        - in: path
          name: testId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestWithQuestions"
    put:
      operationId: updateTest
      tags: [tests]
      summary: Update a test (admin)
      parameters:
        - in: path
          name: testId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Test"
    delete:
      operationId: deleteTest
      tags: [tests]
      summary: Delete a test (admin)
      parameters:
        - in: path
          name: testId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /questions:
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Create a new question (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/{questionId}:
    put:
      operationId: updateQuestion
      tags: [questions]
      summary: Update a question (admin)
      parameters:
        - in: path
          name: questionId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete a question (admin)
      parameters:
        - in: path
          name: questionId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /admin/announcements:
    post:
      operationId: createAnnouncement
      tags: [admin]
      summary: Send announcement to all users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Announcement sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    SendOtpInput:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
    OtpResponse:
      type: object
      required: [message, otpId]
      properties:
        message:
          type: string
        otpId:
          type: string
    VerifyOtpInput:
      type: object
      required: [phone, otp]
      properties:
        phone:
          type: string
        otp:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, phone, name, role, createdAt]
      properties:
        id:
          type: integer
        phone:
          type: string
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [student, admin]
        rank:
          type: ["integer", "null"]
        totalAttempts:
          type: ["integer", "null"]
        avgScore:
          type: ["number", "null"]
        createdAt:
          type: string
    Test:
      type: object
      required: [id, title, subject, type, totalQuestions, duration, maxMarks, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        subject:
          type: string
        type:
          type: string
          enum: [full, physics, chemistry, mathematics]
        totalQuestions:
          type: integer
        duration:
          type: integer
          description: Duration in minutes
        maxMarks:
          type: integer
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, published, archived]
        attemptCount:
          type: ["integer", "null"]
        createdAt:
          type: string
    TestInput:
      type: object
      required: [title, subject, type, duration, maxMarks]
      properties:
        title:
          type: string
        subject:
          type: string
        type:
          type: string
          enum: [full, physics, chemistry, mathematics]
        duration:
          type: integer
        maxMarks:
          type: integer
        description:
          type: string
        status:
          type: string
          enum: [draft, published, archived]
    TestUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, published, archived]
        duration:
          type: integer
    TestWithQuestions:
      type: object
      required: [id, title, subject, type, totalQuestions, duration, maxMarks, status, questions, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        subject:
          type: string
        type:
          type: string
        totalQuestions:
          type: integer
        duration:
          type: integer
        maxMarks:
          type: integer
        description:
          type: ["string", "null"]
        status:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        createdAt:
          type: string
    Question:
      type: object
      required: [id, testId, subject, questionText, questionType, marks, negativeMarks, orderIndex]
      properties:
        id:
          type: integer
        testId:
          type: integer
        subject:
          type: string
          enum: [physics, chemistry, mathematics]
        questionText:
          type: string
        questionType:
          type: string
          enum: [mcq, numerical]
        optionA:
          type: ["string", "null"]
        optionB:
          type: ["string", "null"]
        optionC:
          type: ["string", "null"]
        optionD:
          type: ["string", "null"]
        correctOption:
          type: ["string", "null"]
          description: A, B, C, or D for MCQ
        correctNumerical:
          type: ["number", "null"]
          description: Correct value for numerical questions
        marks:
          type: integer
        negativeMarks:
          type: number
        orderIndex:
          type: integer
        solution:
          type: ["string", "null"]
    QuestionInput:
      type: object
      required: [testId, subject, questionText, questionType, marks, negativeMarks, orderIndex]
      properties:
        testId:
          type: integer
        subject:
          type: string
          enum: [physics, chemistry, mathematics]
        questionText:
          type: string
        questionType:
          type: string
          enum: [mcq, numerical]
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctOption:
          type: string
        correctNumerical:
          type: number
        marks:
          type: integer
        negativeMarks:
          type: number
        orderIndex:
          type: integer
        solution:
          type: string
    QuestionUpdate:
      type: object
      properties:
        questionText:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctOption:
          type: string
        correctNumerical:
          type: number
        solution:
          type: string
...
    UserListResponse:
      type: object
      required: [users, total, page]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
    AnnouncementInput:
      type: object
      required: [title, message]
      properties:
        title:
          type: string
        message:
          type: string