# lib/api-zod/src/generated/types/otpResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JEE Test Mock Free API
 * OpenAPI spec version: 0.1.0
 */

export interface OtpResponse {
  message: string;
  otpId: string;
}