# lib/api-zod/src/generated/types/questionUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JEE Test Mock Free API
 * OpenAPI spec version: 0.1.0
 */

export interface QuestionUpdate {
  questionText?: string;
  optionA?: string;
  optionB?: string;
  optionC?: string;
  optionD?: string;
  correctOption?: string;
  correctNumerical?: number;
  solution?: string;
}