# lib/api-zod/src/generated/types/testInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JEE Test Mock Free API
 * OpenAPI spec version: 0.1.0
 */
import type { TestInputStatus } from './testInputStatus';
import type { TestInputType } from './testInputType';

export interface TestInput {
  title: string;
  subject: string;
  type: TestInputType;
  duration: number;
  maxMarks: number;
  description?: string;
  status?: TestInputStatus;
}