# lib/api-zod/src/generated/types/testUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JEE Test Mock Free API
 * OpenAPI spec version: 0.1.0
 */
import type { TestUpdateStatus } from './testUpdateStatus';

export interface TestUpdate {
  title?: string;
  description?: string;
  status?: TestUpdateStatus;
  duration?: number;
}