# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JEE Test Mock Free API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  phone: string;
  /** @nullable */
  name: string | null;
  /** @nullable */
  email?: string | null;
  role: UserRole;
  /** @nullable */
  rank?: number | null;
  /** @nullable */
  totalAttempts?: number | null;
  /** @nullable */
  avgScore?: number | null;
  createdAt: string;
}